Add [PHPCountable] and [PHPExtension] iterables
In JavaScript, once a object has a length field and has an indexed
getter, it can be used as an array-like by most methods, including the
foreach statement.  Add a [PHPCountable] extended attribute to the
length field in classes where that should be used to provide support
for PHP 'count', and for classes which are countable and have an
indexed getter, add a [PHPExtension] iterable if one was missing.
Generate a default `::getIterator()` method in the helper trait where
possible.

This makes IDLeDOM more closely match the built-in \DOMNamedNodeMap
(for example), which implements Traversable and Countable, even though
the official WebIDL DOM spec doesn't define an iterable on
NamedNodeMap.