Support the [[LegacyNullToEmptyString]] extended attribute
DOMImplementation::createDocument() as well as CharacterData::data and
HTMLElement::innerText allow nulls to be passed to arguments of
DOMString type, in which case the null is expected to be converted to
the empty string.

See also I512bf34d7093bea7ec5d96644a964e76472fb9ba which fixes a place
where Remex was using this legacy functionality and passing null as
the second argument to DOMImplementation::createDocument().

Change-Id: I69f3116a1b5d4700fa6522ec749553c216f72804