New test framework: create explicit superclasses
Instead of fudging the superclasses and creating anonymous class instances,
generate "real" superclass definitions when testing the stubs and helpers.
This helps ensure that we properly inherit and override from the stubs
and helpers and don't end up with duplicate method definitions or
incompatible overrides.

One issue this turned up was an inconsistent definition of the cast()
helper method: the return type hint has been removed so that cast()
is now consistent with PHP 7.2 covariance restrictions on overrides.

Change-Id: Idd8f16cc73cde8079b85864583f890977e731ed0