Wrap `unsigned long` values in the interface as described in interface spec
As described in WebIDL.md, `unsigned long` types in IDL should be
encoded as signed integers in PHP in order to ensure they fit within
the range of a PHP int on 32-bit platforms.  Although this was in the
spec, when we emitted constants in our interfaces we weren't properly
converting the values.

Thanks to Kunal Mehta, who reported this as an issue caught by debian's
autopkgtest CI system.

Bug: T324489
Change-Id: Icf31304350c6a22fde07c68b4604303e0738e6db